+++ packages/falso/src/index.ts
@@ -201,3 +201,5 @@ export {
   incrementalDate,
   IncrementalDateOptions,
 } from './lib/factories/incremental-date';
+export { randAggregation } from './lib/aggregation';
+export { toCollection } from './lib/collection';