+++ packages/falso/src/lib/amount.ts
@@ -2,7 +2,9 @@ import {
   fake,
   FakeOptions,
   getRandomInRange,
+  markRequired,
   RandomInRangeOptions,
+  Return,
 } from './core/core';
 
 export interface RandomAmountOptions extends RandomInRangeOptions, FakeOptions {
@@ -28,12 +30,18 @@ export interface RandomAmountOptions extends RandomInRangeOptions, FakeOptions {
  *
  * @example
  *
- * randAmount({ symbol: '$' }) // $12.52
+ * randAmount({ symbol: '$' }) // '$12.52'
  *
  * @example
  *
  * randAmount({ length: 10 })
  */
+export function randAmount<
+  Options extends markRequired<RandomAmountOptions, 'symbol'>
+>(options: Options): Return<string, Options>;
+export function randAmount<Options extends RandomAmountOptions = never>(
+  options?: Options
+): Return<number, Options>;
 export function randAmount<Options extends RandomAmountOptions = never>(
   options?: Options
 ) {
@@ -44,8 +52,9 @@ export function randAmount<Options extends RandomAmountOptions = never>(
 
   const symbol = options?.symbol ?? '';
 
-  const factory: () => string = () => {
-    return `${symbol}${getRandomInRange(amountOptions)}`;
+  const factory: () => string | number = () => {
+    const num = getRandomInRange(amountOptions);
+    return symbol ? `${symbol}${num}` : num;
   };
 
   return fake(factory, options);