+++ packages/falso/src/lib/number.ts
@@ -2,8 +2,10 @@ import {
   fake,
   FakeOptions,
   getRandomInRange,
+  markRequired,
   RandomInRangeOptions,
 } from './core/core';
+import { isNil } from './core/validators';
 
 export interface RandomNumberOptions extends RandomInRangeOptions, FakeOptions {
   precision?: number;
@@ -20,23 +22,27 @@ export interface RandomNumberOptions extends RandomInRangeOptions, FakeOptions {
  *
  * @example
  *
+ * randNumber({ length: 10 })
+ *
+ * @example
+ *
  * randNumber({ min: 10, max: 1000 }) // default is 'min': 0, 'max': 999_999
  *
  * @example
  *
- * randNumber({ min: 1000, max: 100000, precision: 1000 }) // 67000
+ * randNumber({ min: 0.1, max: 1, fraction: 2 }) // 0.37
  *
  * @example
  *
- * randNumber({ min: 1000, max: 2000, precision: 100 }) // 1200
+ * randNumber({ min: 1000, max: 100000, precision: 1000 }) // 67_000
  *
  * @example
  *
- * randNumber({ min: 1000, max: 2000, precision: 10 }) // 1250
+ * randNumber({ min: 1000, max: 2000, precision: 100 }) // 1_200
  *
  * @example
  *
- * randNumber({ length: 10 })
+ * randNumber({ min: 1000, max: 2000, precision: 10 }) // 1_250
  *
  * @example
  *
@@ -46,16 +52,21 @@ export interface RandomNumberOptions extends RandomInRangeOptions, FakeOptions {
 export function randNumber<Options extends RandomNumberOptions = never>(
   options?: Options
 ) {
-  const o: RandomNumberOptions = {
-    min: options?.min || 0,
-    max: options?.max || 999_999,
+  const normalized: markRequired<RandomNumberOptions, 'min' | 'max'> = {
+    min: isNil(options?.min) ? 0 : options!.min,
+    max: isNil(options?.max) ? 999_999 : options!.max,
     precision: options?.precision,
+    fraction: options?.fraction,
   };
 
   return fake(() => {
-    const num = getRandomInRange(o);
-    if (o.precision !== undefined) {
-      return Math.floor(num / o.precision) * o.precision;
+    if (normalized.min === normalized.max) {
+      return normalized.min;
+    }
+
+    const num = getRandomInRange(normalized);
+    if (normalized.precision !== undefined) {
+      return Math.floor(num / normalized.precision) * normalized.precision;
     }
     return num;
   }, options);